# Improve performance of Path#dirname and Path#extension