@BlobCodes 2021-07-24
Hmm the #extension method seems to fail some specs (appearantly not all Path specs were in path_spec.cr.. hmm)
I'll probably remove it.