@straight-shoota 2021-07-24
This looks great, really. 

To fix `#extension` I think you just need to make sure to find any separators after the last dot. `bytes.rindex(offset: pos) { |byte| byte === '.' || byte.in?(sep) }` should work for that. Then you can check if `bytes.unsafe_fetch(dot_index)` is actually `.` and return empty otherwise.

The specs are in `file_spec` because these methods were originally implemented as class methods of `File`. We just moved the implementation to `Path` when that was introduced, and I kept the original specs exactly like they were to demonstrate backwards-compatibility. I think we should better move them to `path_spec`, though (that should better be a separate PR).

Your code comments in `#dirname` are really great and explain what's going on really well. `#extname` on the other hand could use a couple more on the special cases handling. The previous ones could just carry over.