@BlobCodes 2021-07-24
It seems like Path#extension now properly works.
It is not as fast as before, but still an improvement.